底盘功率待完成

+++ Chassis_2023/File/API/can_protocol.h
@@ -6,7 +6,7 @@
 
 s32 GetEncoderNumber(u8* pData);
 s32 GetSpeed(u8* pData);
-void AbsEncoderProcess(ST_ENCODER* pEncoder, s32 value);
+void AbsEncoderProcess(ST_ENCODER* pEncoder, float value);
 void CAN_SendData(CAN_TypeDef* CANx,u32 id,s16 Data1,s16 Data2,s16 Data3,s16 Data4);
 void CAN1_Protocol(void);
 void CAN2_Protocol(void);
